+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 - Removed Injecter.WPF inheritable classes
 - When using the InjectScoped property you can get the create scope via the IScopeStore interface
-- Updated docs
+- Removed Injecter.UWP inheritable classes
+- Added attached property injection to Injecter.UWP